jobs and other stuff.
git-svn-id: http://svn.netlabs.org/repos/kbuild/trunk@1207 8c291193-6614-0410-a08d-adc6e07170bb

+++ src/kash/mystring.c
@@ -78,7 +78,7 @@ char nullstr[1];		/* zero length string */
  */
 
 void
-scopyn(const char *from, char *to, int size)
+scopyn(const char *from, char *to, ssize_t size)
 {
 
 	while (--size > 0) {
